got usecontext working
