+++ src/App.jsx
@@ -1,6 +1,7 @@
 import './styles/styles.scss';
 import AppRouter from './routers/AppRouter';
 
+
 function App() {
 
   return (
