+++ src/pages/PageFavourites.jsx
@@ -1,19 +1,22 @@
 import { useContext, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-
-
-// Find way to set global state
-// pull in global data of favourite movies
-
+import { FavContext } from '../context/favContext';
 
 function PageFavourites() {
 
-    const favsArray = JSON.parse(localStorage.getItem('movieData'));
+    const { favs, setFavs } = useContext(FavContext);
+
+    useEffect(() => {
+        const storedFavs = JSON.parse(localStorage.getItem('movieData'));
+        if (storedFavs) {
+            setFavs(storedFavs);
+        }
+    }, [setFavs]);
 
     return (
         <section>
             <ul>
-                {favsArray && favsArray.map(item => {
+                {favs && favs.map(item => {
                     return (
                         <li key={item.id + "fav"} >
                             <Link to={`./details/${item.id}`}><img src={`https://image.tmdb.org/t/p/w342${item.poster_path}`} alt={item.title} /></Link>
@@ -28,14 +31,6 @@ function PageFavourites() {
                     )
                 })}
             </ul>
-
-
-
-
-            {/* <h2>Your Favourites</h2>
-            <h3>{favMsg}</h3>
-            <h3>{favTitle}</h3>
-            <img src={`https://image.tmdb.org/t/p/w342${favImg}`} alt={favTitle} /> */}
         </section>
     )
 }
